Properly handle EINTR in POSIX UDP and TCP transport profiles
Blocking I/O functions, like connect(), poll(), send() and recv(), will return
-1 with errno == EINTR if the process/thread is interrupted by a signal. This
is not an I/O error, but generally means the operations should be restarted.

This commit implements the auto-restart behavior, which allows the transport
profiles to be used in environments that generate a lot of signals, like the
FreeRTOS simulator for Linux.

Signed-off-by: J. S. Seldenthuis <jseldenthuis@lely.com>